refactor: sell input row refactor
