+++ apps/web/src/app/[locale]/shop/_auction/SellSection/SellInputRow.tsx
@@ -12,8 +12,8 @@ import { useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
 import { useRegisterProduct } from '@/apis/auctions/useRegisterProduct';
-import { useGetAllPersona } from '@/hooks/query/render/useGetAllPersona';
-import { ANIMAL_TIER_TEXT_MAP, getAnimalTierInfo } from '@/utils/animals';
+import { useGetPersonaTier } from '@/hooks/persona/useGetPersonaDropRate';
+import { ANIMAL_TIER_TEXT_MAP } from '@/utils/animals';
 import { getPersonaImage } from '@/utils/image';
 
 import { rowStyle, ShopTableMobileRow } from '../../_common/ShopTableMobileRow';
@@ -22,7 +22,7 @@ import { tableCss, theadCss } from '../table.styles';
 const MAX_PRICE = 100_000_000;
 
 interface Props {
-  item: Persona | null;
+  item: Persona;
   initPersona: () => void;
 }
 
@@ -33,10 +33,7 @@ function SellInputRow({ item, initPersona }: Props) {
 
   const queryClient = useQueryClient();
 
-  const {
-    data: { personas },
-  } = useGetAllPersona();
-  const currentPersona = personas.find((persona) => persona.type === item?.type);
+  const personaTier = useGetPersonaTier(item.type);
 
   const { mutate } = useRegisterProduct({
     onSuccess: () => {
@@ -64,7 +61,6 @@ function SellInputRow({ item, initPersona }: Props) {
   };
 
   if (isMobile) {
-    if (!item) return null;
     return (
       <ShopTableMobileRow
         personaType={item.type}
@@ -90,13 +86,13 @@ function SellInputRow({ item, initPersona }: Props) {
       </div>
 
       <div className={cx(rowStyle, 'row')}>
-        {item && currentPersona && (
+        {item && personaTier && (
           <>
             <div>
               <img src={getPersonaImage(item.type)} alt={item.type} width={60} height={67} />
             </div>
             <div>{snakeToTitleCase(item.type)}</div>
-            <div>{ANIMAL_TIER_TEXT_MAP[getAnimalTierInfo(Number(currentPersona.dropRate.replace('%', '')))]}</div>
+            <div>{ANIMAL_TIER_TEXT_MAP[personaTier]}</div>
             <div>{item.level}</div>
             <div>
               <input
