--- apps/web/src/app/[locale]/shop/_auction/SellSection/index.ts
@@ -1,1 +1,0 @@
-export { default as SellSection } from './SellSection';
